Add road id and proper timestamp values. Bug fixes for file parsing and mappings between pipeline stages

+++ HiveWays.Core/HiveWays.Domain/Models/DataPoint.cs
@@ -5,7 +5,7 @@ namespace HiveWays.Domain.Models;
 public class DataPoint : IIdentifiable
 {
     [Name("#time")]
-    public int TimeOffsetSeconds { get; set; }
+    public double TimeOffsetSeconds { get; set; }
 
     [Name("id")]
     public int Id { get; set; }
